article_factory/requirements.txt
google-adk[all]
google-genai
deprecated
